[Feat#1] 모바일 헤더 초기
