+++ src/components/common/Header/ClientHeader.tsx
@@ -1,4 +1,3 @@
-// ClientHeader.tsx
 "use client";
 
 import Header from "@/components/common/Header";
